Don't perform teardown when test data is not available since setup was not completed either
git-svn-id: http://svn.codehaus.org/geoserver/trunk/geoserver@8523 ef1d6e69-97e5-0310-af46-8a06194da32a

Former-commit-id: 71754c381a9e50227b7fc879d0f43a66f21081ca [formerly eabfb0d53c55811330725ed37b62d16dda215ee3]
Former-commit-id: 4450e770590d0dd6e720f33ad4d7b69a0a595d9f

+++ main/src/test/java/org/geoserver/test/GeoServerAbstractTestSupport.java
@@ -208,20 +208,23 @@ public abstract class GeoServerAbstractTestSupport extends TestCase {
     protected void tearDown() throws Exception {
         super.tearDown();
         
-        //kill the context
-        applicationContext.destroy();
-        applicationContext = null;
-
-        if(isMemoryCleanRequired())
-            System.gc(); System.runFinalization();
-        
-        if(getTestData() != null) {
-            // this cleans up the data directory static loader, if we don't the next test
-            // will keep on running on the current data dir
-            GeoserverDataDirectory.destroy();
-            getTestData().tearDown();
+        if(getTestData().isTestDataAvailable()) {
+            //kill the context
+            applicationContext.destroy();
+            applicationContext = null;
+    
+            if(isMemoryCleanRequired())
+                System.gc(); System.runFinalization();
+            
+            if(getTestData() != null) {
+                // this cleans up the data directory static loader, if we don't the next test
+                // will keep on running on the current data dir
+                GeoserverDataDirectory.destroy();
+                getTestData().tearDown();
+            }
         }
     }
+        
 
     /**
      * Accessor for global catalog instance from the test application context.
